requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
passlib==1.7.4
python-jose==3.5.0
pydantic==2.11.7
email-validator==2.3.0
python-multipart==0.0.12
pymysql==1.1.2
python-dotenv==1.1.1
bcrypt==4.3.0
pyjwt==2.10.1